ongoing code layout impr
